refactor: restructure project, rename files, and add styles for main files (App.tsx, home.tsx, index.tsx, layout.tsx)
After applying this commit, the project structure will be reorganized, files will be renamed, and styles will be added for the main files: App.tsx, home.tsx, index.tsx, and layout.tsx.
