+++ src/layouts/footer/index.tsx
