--- src/pages/home/home.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Home: React.FC = () => {
-	return <h2>Главная страница</h2>;
-};
-
-export default Home;
